Refactored Dynamic Dropdown
- moved dropdowns into its own file under data/dropdowns
- moved keyCompare into its own file under data/dropdowns
