+++ src/components/DynamicDropdown.js
@@ -1,81 +1,6 @@
-// dropdown data imports (for *really big* categories)
-import country from '../data/dropdowns/country';
-import genres from '../data/dropdowns/genres';
-import language from '../data/dropdowns/language';
-import network from '../data/dropdowns/network';
-import webChannel from '../data/dropdowns/webChannel';
+import dropdowns from '../data/dropdowns/dropdowns';
 
-const dropdowns = {
-
-  status: [
-    "Running",
-    "Ended",
-    "To Be Determined",
-    "In Development",
-  ],
-
-  type: [
-    "Scripted",
-    "Animation",
-    "Reality",
-    "Talk Show",
-    "Documentary",
-    "Game Show",
-    "News",
-    "Sports",
-    "Variety",
-    "Award Show",
-    "Panel Show",
-  ],
-
-  genres: genres,
-
-  language: language,
-
-  country: country,
-
-  // has optgroups
-  network: network,
-
-  // has optgroups
-  webChannel: webChannel,
-
-  // has separate str and val properties
-  runtime: [
-    {
-      str: "30 minutes or less",
-      val: {
-        max: 30,
-        min: 0,
-      }
-    },
-    {
-      str: "30 to 60 minutes",
-      val: {
-        max: 60,
-        min: 30,
-      }
-    },
-    {
-      str: "Over 60 minutes",
-      val: {
-        max: Infinity,
-        min: 60,
-      }
-    },
-  ],
-
-  rating: [
-    2,
-    3,
-    4,
-    5,
-    6,
-    7,
-    8,
-    9,
-  ],
-}
+const categories = Object.keys(dropdowns);
 
 // the default value to be displayed on all dropdowns
 const DEFAULT_OPTION = "";
@@ -162,76 +87,5 @@ function DynamicDropdown({ category, state, setState }) {
   );
 };
 
-/**
- * helper function holding the logic of show/filter key comparisons
- * @param {string} key a key from the dropdowns object
- * @param {object} show a tv show object from the TVmaze API
- * @param {object} filters the "filters" reduced state from UserInput
- * @returns true if show[key] matches the constraints of filters[key]
- */
-const keyCompare = (key, show, filters) => {
-
-  // if this key is unspecified in filter, 
-  // evaluate as true
-  if (!filters[key]) {
-    return true;
-  }
-
-  // if this key is specified in filter but not on the show object, 
-  // evaluate as false
-  if (!show[key]) {
-    return false;
-  }
-
-  let match = true;
-
-  // check whether this key requires a special testing procedure
-  switch (key) {
-
-    default:
-      match = (show[key] === filters[key]);
-      break;
-
-    // special cases
-
-    // genres must *include* selected genre
-    case 'genres':
-      match = (show.genres.includes(filters.genres));
-      break;
-
-    // runtime must be *between* selected bounds
-    case 'runtime':
-      const { runtime } = show;
-      const { min, max } = filters.runtime;
-      match = (
-        min <= runtime && runtime <= max
-      );
-      break;
-
-    // rating must be selected figure *or higher*
-    case 'rating':
-      match = (show.rating >= filters.rating);
-      break;
-
-    // test network/webchannel *name*
-    case 'network':
-    case 'webChannel':
-      match = (show[key].name === filters[key]);
-      break;
-
-    // shows have no "country" property;
-    // 1. check for network or webchannel
-    // 2. test (network/webchannel).country.name
-    case 'country':
-      const country =
-        (show.network || show.webChannel).country.name;
-
-      match = (country === filters.country);
-      break;
-  }
-
-  return match;
-}
-
 export default DynamicDropdown;
-export { dropdowns, keyCompare };
+export { categories };
